requirements.txt
discord.py
requests
pandas
pandas-ta
ta
